feat(email): improve thread card layout and add action buttons
- Remove subject from individual thread cards (already shown in header)
- Move action buttons (star, clock, trash, ellipsis) from cards to thread header
- Add Reply, Reply All, Forward buttons to bottom of thread cards
- Show date and time instead of just date in thread cards
- Remove metadata line (participants, messages, date) from thread header
- Clean up unused imports and handler functions from thread cards
