+++ frontend/components/email/email-thread-card.tsx
@@ -1,14 +1,9 @@
 import { Button } from '@/components/ui/button';
-import { DropdownMenu, DropdownMenuContent, DropdownMenuItem, DropdownMenuTrigger } from '@/components/ui/dropdown-menu';
-import { useShipmentDetection } from '@/hooks/use-shipment-detection';
-import { shipmentsClient } from '@/lib/shipments-client';
 import { EmailMessage } from '@/types/office-service';
 import DOMPurify from 'dompurify';
-import { Archive, Clock, Download, MoreHorizontal, Reply, Star, Trash2, Wand2 } from 'lucide-react';
-import React, { useState } from 'react';
-import { toast } from 'sonner';
+import { Forward, Reply, ReplyAll } from 'lucide-react';
+import React from 'react';
 import AISummary from './ai-summary';
-import TrackShipmentModal, { PackageFormData } from './track-shipment-modal';
 
 // Configure DOMPurify for email content
 const emailSanitizeConfig = {
@@ -86,7 +81,7 @@ const formatEmailDate = (dateString: string): string => {
     const today = new Date(now.getFullYear(), now.getMonth(), now.getDate());
     const emailDay = new Date(emailDate.getFullYear(), emailDate.getMonth(), emailDate.getDate());
 
-    // If email was sent today, show time
+    // If email was sent today, show time only
     if (emailDay.getTime() === today.getTime()) {
         return emailDate.toLocaleTimeString([], {
             hour: 'numeric',
@@ -95,10 +90,14 @@ const formatEmailDate = (dateString: string): string => {
         });
     }
 
-    // Otherwise show month and day
+    // Otherwise show date and time
     return emailDate.toLocaleDateString([], {
         month: 'short',
         day: 'numeric'
+    }) + ' ' + emailDate.toLocaleTimeString([], {
+        hour: 'numeric',
+        minute: '2-digit',
+        hour12: true
     });
 };
 
@@ -141,91 +140,19 @@ const EmailThreadCard: React.FC<EmailThreadCardProps> = ({
     onSelect,
     inlineAvatar = false
 }) => {
-    const [isModalOpen, setIsModalOpen] = useState(false);
-    const [isStarred, setIsStarred] = useState(false);
-    const shipmentDetection = useShipmentDetection(email);
-
-    const handleDownload = async () => {
-        try {
-            const testData = {
-                provider: email.provider,
-                date: email.date,
-                subject: email.subject,
-                sender: email.from_address?.email || '',
-                body_data: {
-                    contentType: email.body_html ? "HTML" : "Text",
-                    content: email.body_html || email.body_text || ""
-                }
-            };
-
-            const blob = new Blob([JSON.stringify(testData, null, 2)], {
-                type: 'application/json'
-            });
-            const url = URL.createObjectURL(blob);
-            const a = document.createElement('a');
-            a.href = url;
-            a.download = `email_test_${email.id}_${Date.now()}.json`;
-            document.body.appendChild(a);
-            a.click();
-            document.body.removeChild(a);
-            URL.revokeObjectURL(url);
-
-        } catch (error) {
-            console.error('Failed to download email:', error);
-            alert('Failed to download email. Please try again.');
-        }
-    };
-
-    const handleTrackShipment = () => {
-        setIsModalOpen(true);
-    };
-
-    const handleModalClose = () => {
-        setIsModalOpen(false);
-    };
-
-    const handleTrackShipmentSubmit = async (packageData: PackageFormData) => {
-        try {
-            const packageDataWithEmail = {
-                ...packageData,
-                email_message_id: email.id,
-            };
-
-            const createdPackage = await shipmentsClient.createPackage(packageDataWithEmail);
-
-            console.log('Package created successfully:', createdPackage);
-
-            toast.success(`Successfully started tracking package ${packageData.tracking_number}`);
-
-        } catch (error) {
-            console.error('Failed to create package:', error);
-            throw error;
-        }
-    };
-
-    const handleStarToggle = () => {
-        setIsStarred(!isStarred);
-        // TODO: Implement actual star functionality
-    };
-
     const handleReply = () => {
         // TODO: Implement reply functionality
         console.log('Reply to email:', email.id);
     };
 
-    const handleArchive = () => {
-        // TODO: Implement archive functionality
-        console.log('Archive email:', email.id);
+    const handleReplyAll = () => {
+        // TODO: Implement reply all functionality
+        console.log('Reply all to email:', email.id);
     };
 
-    const handleSnooze = () => {
-        // TODO: Implement snooze functionality
-        console.log('Snooze email:', email.id);
-    };
-
-    const handleDelete = () => {
-        // TODO: Implement delete functionality
-        console.log('Delete email:', email.id);
+    const handleForward = () => {
+        // TODO: Implement forward functionality
+        console.log('Forward email:', email.id);
     };
 
     const senderName = email.from_address?.name || email.from_address?.email || 'Unknown';
@@ -272,108 +199,15 @@ const EmailThreadCard: React.FC<EmailThreadCardProps> = ({
                                     {email.has_attachments && (
                                         <span className="text-gray-400">📎</span>
                                     )}
-                                    {shipmentDetection.isShipmentEmail && (
-                                        <span className="px-2 py-1 bg-green-100 text-green-700 rounded text-xs">
-                                            📦 Shipment
-                                        </span>
-                                    )}
-                                </div>
-                                <div className={`text-lg mb-2 ${isUnread ? 'font-semibold' : ''}`}>
-                                    {email.subject || '(No subject)'}
                                 </div>
                                 <div className="text-sm text-gray-600">
                                     To: {email.to_addresses.map(addr => addr.name || addr.email).join(', ')}
                                 </div>
                             </div>
 
-                            {/* Date and actions */}
-                            <div className="flex-shrink-0 flex flex-col items-end gap-2">
+                            {/* Date */}
+                            <div className="flex-shrink-0">
                                 <span className="text-sm text-gray-500">{formattedDate}</span>
-
-                                {/* Action buttons - always visible */}
-                                <div className="flex items-center gap-1">
-                                    <Button
-                                        variant="ghost"
-                                        size="sm"
-                                        className="h-8 w-8 p-0"
-                                        onClick={(e) => {
-                                            e.stopPropagation();
-                                            handleStarToggle();
-                                        }}
-                                        title="Star"
-                                    >
-                                        <Star className={`h-4 w-4 ${isStarred ? 'fill-current text-yellow-500' : ''}`} />
-                                    </Button>
-                                    <Button
-                                        variant="ghost"
-                                        size="sm"
-                                        className="h-8 w-8 p-0"
-                                        onClick={(e) => {
-                                            e.stopPropagation();
-                                            handleSnooze();
-                                        }}
-                                        title="Snooze"
-                                    >
-                                        <Clock className="h-4 w-4" />
-                                    </Button>
-                                    <Button
-                                        variant="ghost"
-                                        size="sm"
-                                        className="h-8 w-8 p-0"
-                                        onClick={(e) => {
-                                            e.stopPropagation();
-                                            handleDelete();
-                                        }}
-                                        title="Delete"
-                                    >
-                                        <Trash2 className="h-4 w-4" />
-                                    </Button>
-                                    <DropdownMenu>
-                                        <DropdownMenuTrigger asChild>
-                                            <Button
-                                                variant="ghost"
-                                                size="sm"
-                                                className="h-8 w-8 p-0"
-                                                onClick={(e) => e.stopPropagation()}
-                                                title="More actions"
-                                            >
-                                                <MoreHorizontal className="h-4 w-4" />
-                                            </Button>
-                                        </DropdownMenuTrigger>
-                                        <DropdownMenuContent align="end">
-                                            <DropdownMenuItem onClick={(e) => {
-                                                e.stopPropagation();
-                                                handleReply();
-                                            }}>
-                                                <Reply className="h-4 w-4 mr-2" />
-                                                Reply
-                                            </DropdownMenuItem>
-                                            <DropdownMenuItem onClick={(e) => {
-                                                e.stopPropagation();
-                                                handleArchive();
-                                            }}>
-                                                <Archive className="h-4 w-4 mr-2" />
-                                                Archive
-                                            </DropdownMenuItem>
-                                            <DropdownMenuItem onClick={(e) => {
-                                                e.stopPropagation();
-                                                handleDownload();
-                                            }}>
-                                                <Download className="h-4 w-4 mr-2" />
-                                                Download
-                                            </DropdownMenuItem>
-                                            {shipmentDetection.isShipmentEmail && (
-                                                <DropdownMenuItem onClick={(e) => {
-                                                    e.stopPropagation();
-                                                    handleTrackShipment();
-                                                }}>
-                                                    <Wand2 className="h-4 w-4 mr-2" />
-                                                    Track Shipment
-                                                </DropdownMenuItem>
-                                            )}
-                                        </DropdownMenuContent>
-                                    </DropdownMenu>
-                                </div>
                             </div>
                         </div>
                     </div>
@@ -422,20 +256,25 @@ const EmailThreadCard: React.FC<EmailThreadCardProps> = ({
                         size="sm"
                         onClick={(e) => {
                             e.stopPropagation();
-                            handleArchive();
+                            handleReplyAll();
                         }}
                     >
-                        <Archive className="h-4 w-4 mr-1" />
-                        Archive
+                        <ReplyAll className="h-4 w-4 mr-1" />
+                        Reply All
+                    </Button>
+                    <Button
+                        variant="outline"
+                        size="sm"
+                        onClick={(e) => {
+                            e.stopPropagation();
+                            handleForward();
+                        }}
+                    >
+                        <Forward className="h-4 w-4 mr-1" />
+                        Forward
                     </Button>
                 </div>
             </div>
-            <TrackShipmentModal
-                isOpen={isModalOpen}
-                onClose={handleModalClose}
-                email={email}
-                onTrackShipment={handleTrackShipmentSubmit}
-            />
         </>
     );
 };
